Switch to using stdio_raw() instead of stdio() for the test-tool
stdio() used to be unbuffered, but recently became buffered. So,
calling the method multiple times resulted in parts of the STDIN
stream being thrown away. Using the unbuffered stdin_raw() fixes
this. It would also be possible to save the result of stdio() and
not call it multiple times, but, that can be saved for a later
change.