aes::ctr().process() would infinite loop
The SynchronousStreamCipher implementation on boxed SynchronousStreamCiphers
just calls the process method on the underlying SynchronousStreamCipher
trait object. In older Rust versions, I believed that this worked. It
doesn't work now - it just calls itself and infinite loops until the
stack is oveflowed. Unfortunately, there wasn't a test that caught this.
This fixes the underlying problem and adds a test.