Aessafe: Re-write the change_basis() function
The change_basis() function take a matrix as a parameter
for the conversation to perform. That function is removed
and replaced with 4 different functions (1 for each matrix)
that have been unrolled and optimized. The matrixes can then
be removed. This improves performance quite a bit for both
Aessafe implementations, especially the X8 version.