Add support for XChaCha20.
XChaCha20 is the extended 192 bit nonce variant of the ChaCha20 stream
cipher.  Like the current existing XSalsa20 implementation, only 256
bit keys are supported.

Note: As there are no existing test vectors for this algorithm, so a
simple test was generated using Andrew Moon's chacha-opt library, using
the same key and nonce as the XSalsa20 test case.