Use std::simd for Salsa20.
It kind of uses macros like the branch sitting in pull request #230, but
some things are still left as functions.  Macroing everything reduced
performance on my test setup, but I haven't examined it in depth.

The state structure is harder to untangle because the layout that's
required for vectorized round functions looks nothing like the "natural"
Salsa20 matrix, but I tried to use variable names and comments to
hightlight what lives where (ChaCha20 is a lot nicer to vectorize since
some of the changes were motivated by such).

After:
 * salsa20::bench::salsa20_64k  242093 ns/iter (+/- 1784) = 270 MB/s