Fix the implementation of reset() for Blake2b
* Nothing stops you from createing a keyed Blake2b Digest and then
  calling the reset method on the Digest. However, the reset()
  implementation would then fail to reset the key - this is surprising
  behavior.
* Both the Digest and Mac reset() implementations would always use the
  default_params() regardless of the parameters originally used to
  create the Blake2b instance.

So, the Blake2b structure is extended to save a copy of the params used
to create the Blake2b instance. The Mac and Digest reset() methods are
replaced with a reset() method on Blake2b itself that resets the
instance using the stored params and re-applies the key if necessary.