Remove rustc-serialize usage
Now that rustc-serialize has been marked as deprecated, it seems prudent
to drop it's usage and replace the functionality with other crates.

Within this project, it was only used for base64 and hex encoding. Also,
the hex encoding is for testing purposes. Other crates exist for those
purposes, making this change quite easy.